# openapi.yaml
openapi: 3.1.0
info:
  title: Google Calendar API Integration
  description: API for managing Google Calendar events, including creating, fetching, and updating events.
  version: 1.0.0
servers:
  - url: https://4b7b-2407-d000-1a-4cf4-bdd9-df33-df5c-a993.ngrok-free.app/
    description: Local development server
paths:
  /privacy:
    get:
      summary: Privacy Policy
      description: Returns the privacy policy text as an HTML page.
      operationId: getPrivacyPolicy
      responses:
        200:
          description: Privacy policy HTML content.
          content:
            text/html:
              schema:
                type: string

  /startAuth:
    get:
      summary: Start OAuth Authentication
      description: Initiates the OAuth authentication flow.
      operationId: startAuthFlow
      responses:
        200:
          description: Authentication URL.
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    format: uri

  /handleAuth:
    get:
      summary: Handle OAuth Callback
      description: Completes the OAuth authentication flow and stores user credentials.
      operationId: handleAuthCallback
      responses:
        200:
          description: Authentication successful message.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /createEvent:
    post:
      summary: Create a Google Calendar Event
      description: Creates a new event in the user's Google Calendar.
      operationId: createCalendarEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the event.
                  example: "Meeting with Team"
                start_time:
                  type: string
                  format: date-time
                  description: Start time of the event in ISO 8601 format.
                  example: "2024-12-05T10:00:00Z"
                end_time:
                  type: string
                  format: date-time
                  description: End time of the event in ISO 8601 format.
                  example: "2024-12-05T11:00:00Z"
                location:
                  type: string
                  description: Location of the event.
                  example: "Conference Room"
                description:
                  type: string
                  description: Description of the event.
                  example: "Discussing project progress"
      responses:
        200:
          description: Event created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event_link:
                    type: string
                    format: uri
        400:
          description: Missing or invalid input.
        500:
          description: Internal server error.

  /getEvents:
    get:
      summary: Get Events on a Specific Date
      description: Fetches all events scheduled for a specific date from Google Calendar.
      operationId: getEventsByDate
      parameters:
        - name: date
          in: query
          required: true
          description: Date to fetch events for (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of events for the specified date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        summary:
                          type: string
                        start:
                          type: object
                          properties:
                            dateTime:
                              type: string
                              format: date-time
                        end:
                          type: object
                          properties:
                            dateTime:
                              type: string
                              format: date-time
        400:
          description: Invalid date parameter.
        500:
          description: Internal server error.

  /updateEvent:
    put:
      summary: Update a Google Calendar Event
      description: Updates the details of an existing event by its ID.
      operationId: updateCalendarEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: string
                  description: The ID of the event to update.
                  example: "abc123"
                title:
                  type: string
                  description: The new title of the event.
                start_time:
                  type: string
                  format: date-time
                  description: The new start time of the event.
                end_time:
                  type: string
                  format: date-time
                  description: The new end time of the event.
                location:
                  type: string
                  description: The new location of the event.
                description:
                  type: string
                  description: The new description of the event.
      responses:
        200:
          description: Event updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    type: object
        400:
          description: Missing or invalid event ID.
        500:
          description: Internal server error.